wip; looks like the db client has a lot of deltas
