+++ package.json
@@ -38,9 +38,14 @@
     "@architect/eslint-config": "2.0.1",
     "@architect/req-res-fixtures": "git+https://github.com/architect/req-res-fixtures.git",
     "@architect/sandbox": "^5.3.4",
+    "@aws-sdk/client-apigatewaymanagementapi": "^3.213.0",
+    "@aws-sdk/client-dynamodb": "^3.213.0",
+    "@aws-sdk/client-sns": "^3.213.0",
+    "@aws-sdk/client-sqs": "^3.213.0",
+    "@aws-sdk/client-ssm": "^3.213.0",
+    "@aws-sdk/lib-dynamodb": "^3.214.0",
     "@types/aws-lambda": "^8.10.108",
     "@types/node": "^18.11.9",
-    "aws-sdk": "2.1055.0",
     "cross-env": "~7.0.3",
     "eslint": "~8.27.0",
     "mock-fs": "~5.2.0",
