+++ src/tables/dynamo.js
@@ -1,25 +1,24 @@
+let dynamo = require('@aws-sdk/client-dynamodb')
+let docclient = require('@aws-sdk/lib-dynamodb')
 let https = require('https')
-let getPorts = require('../lib/get-ports')
+let getPorts = require('../_get-ports')
 let db, doc
 
 /**
  * Instantiates Dynamo service interfaces
  */
 function getDynamo (type, callback) {
-  if (!type) throw ReferenceError('Must supply Dynamo service interface type')
-
-  // We really only want to load aws-sdk if absolutely necessary
-  // eslint-disable-next-line
-  let dynamo = require('aws-sdk/clients/dynamodb')
-
+  if (!type)
+    throw ReferenceError('Must supply Dynamo service interface type')
   let { ARC_ENV, ARC_LOCAL, AWS_REGION } = process.env
   let local = ARC_ENV === 'testing' || ARC_LOCAL
-  let DB = dynamo
-  let Doc = dynamo.DocumentClient
+  let DB = dynamo.DynamoDB
+  let Doc = docclient.DynamoDBDocument
 
   if (db && type === 'db') {
     return callback(null, db)
   }
+
   if (doc && type === 'doc') {
     return callback(null, doc)
   }
@@ -35,14 +34,9 @@ function getDynamo (type, callback) {
       httpOptions: { agent }
     }
 
-    if (type === 'db') {
-      db = new DB(config)
-      return callback(null, db)
-    }
-    if (type === 'doc') {
-      doc = new Doc(config)
-      return callback(null, doc)
-    }
+    db = new DB(config)
+    doc = Doc.from(db)
+    return callback(null, type === 'db' ? db : doc)
   }
   else {
     getPorts((err, ports) => {
@@ -56,20 +50,15 @@ function getDynamo (type, callback) {
           endpoint: `http://localhost:${port}`,
           region: AWS_REGION || 'us-west-2' // Do not assume region is set!
         }
-        if (type === 'db') {
-          db = new DB(config)
-          return callback(null, db)
-        }
-        if (type === 'doc') {
-          doc = new Doc(config)
-          return callback(null, doc)
-        }
+        db = new DB(config)
+        doc = Doc.from(db)
+        return callback(null, type === 'db' ? db : doc)
       }
     })
   }
 }
 
 module.exports = {
-  db: getDynamo.bind({}, 'db'),
   doc: getDynamo.bind({}, 'doc'),
+  db: getDynamo.bind({}, 'db'),
 }
