+++ src/ws/index.js
@@ -1,9 +1,9 @@
+const { ApiGatewayManagementApi
+} = require('@aws-sdk/client-apigatewaymanagementapi')
+
 let _api
 function instantiateAPI () {
   if (_api) return
-  // We really only want to load aws-sdk if absolutely necessary
-  // eslint-disable-next-line
-  let ApiGatewayManagementApi = require('aws-sdk/clients/apigatewaymanagementapi')
   let {
     ARC_ENV,
     ARC_LOCAL,
