+++ test/integration/http-session-dynamo-test.js
@@ -118,3 +118,4 @@ test('Teardown', async t => {
   let result = await sandbox.end()
   t.equal(result, 'Sandbox successfully shut down', result)
 })
+
